requirements.txt
requests
mutagen
pyyaml
yt-dlp
